requirements.txt
# RLF Prediction - Online KNN and Embedding Model
# Python 3.8+

# Core ML/DL
tensorflow>=2.10.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Hyperparameter Optimization
optuna>=3.0.0

# Visualization
plotly>=5.10.0
matplotlib>=3.5.0

# Utilities
python-dotenv>=0.19.0
